+++ tests/e2e/cypress/integration/dash_test.ts
@@ -8,16 +8,16 @@ describe('dash basic', () => {
     });
 
     it('can get cell', () => {
-        DashTable.getCell(0, 2).click();
-        DashTable.getCell(0, 2).within(() => cy.get('input').should('have.value', '0'));
+        DashTable.getCell(0, 0).click();
+        DashTable.getCell(0, 0).within(() => cy.get('input').should('have.value', '0'));
 
         cy.get('button.next-page').click();
-        DashTable.getCell(0, 2).within(() => cy.get('input').should('have.value', '250'));
+        DashTable.getCell(0, 0).within(() => cy.get('input').should('have.value', '250'));
     });
 
     it('cell click selects all text', () => {
-        DashTable.getCell(0, 3).click();
-        DashTable.getCell(0, 3).within(() =>
+        DashTable.getCell(0, 1).click();
+        DashTable.getCell(0, 1).within(() =>
             cy.get('input').then($inputs => {
                 const $input = $inputs[0];
 
@@ -29,7 +29,7 @@ describe('dash basic', () => {
 
     // https://github.com/plotly/dash-table/issues/50
     it('can edit last and update dataframe on "enter"', () => {
-        DashTable.getCell(249, 2).click();
+        DashTable.getCell(249, 0).click();
         DOM.focused.then($input => {
             const initialValue = $input.val();
 
@@ -42,12 +42,12 @@ describe('dash basic', () => {
     });
 
     it('can edit last and update dataframe when clicking outside of cell', () => {
-        DashTable.getCell(249, 2).click();
+        DashTable.getCell(249, 0).click();
         DOM.focused.then($input => {
             const initialValue = $input.val();
 
             DOM.focused.type(`abc`);
-            DashTable.getCell(248, 2).click();
+            DashTable.getCell(248, 0).click();
 
             cy.get('#container').should($container => {
                 expect($container.first()[0].innerText).to.equal(`[249][0] = ${initialValue} -> abc`);
@@ -56,7 +56,7 @@ describe('dash basic', () => {
     });
 
     it('can get cell with double click', () => {
-        DashTable.getCell(3, 3).within(() => cy.get('div').dblclick());
-        DashTable.getCell(3, 3).should('have.class', 'focused');
+        DashTable.getCell(3, 1).within(() => cy.get('div').dblclick());
+        DashTable.getCell(3, 1).should('have.class', 'focused');
     });
 });