+++ tests/e2e/cypress/integration/select_row_test.ts
@@ -1,7 +1,7 @@
 import DashTable from 'cypress/DashTable';
 
 describe('delete', () => {
-    describe('be', () => {
+    describe('be pagination & sort', () => {
         beforeEach(() => cy.visit('http://localhost:8081'));
 
         it('can select row', () => {
@@ -10,7 +10,7 @@ describe('delete', () => {
         });
 
         it('can select row when sorted', () => {
-            cy.get('tr th.column-2 .sort').click({ force: true });
+            cy.get('tr th.column-0 .sort').click({ force: true });
             DashTable.getSelect(0).within(() => cy.get('input').click());
             DashTable.getSelect(0).within(() => cy.get('input').should('be.checked'));
         });
@@ -18,12 +18,12 @@ describe('delete', () => {
         it('select, sort, new row is not selected', () => {
             DashTable.getSelect(0).within(() => cy.get('input').click());
             DashTable.getSelect(0).within(() => cy.get('input').should('be.checked'));
-            cy.get('tr th.column-2 .sort').click({ force: true });
+            cy.get('tr th.column-0 .sort').click({ force: true });
             DashTable.getSelect(0).within(() => cy.get('input').should('not.be.checked'));
         });
     });
 
-    describe('fe', () => {
+    describe('fe pagination & sort', () => {
         beforeEach(() => cy.visit('http://localhost:8080'));
 
         it('can select row', () => {
@@ -32,7 +32,7 @@ describe('delete', () => {
         });
 
         it('can select row when sorted', () => {
-            cy.get('tr th.column-2 .sort').click({ force: true });
+            cy.get('tr th.column-0 .sort').click({ force: true });
             DashTable.getSelect(0).within(() => cy.get('input').click());
             DashTable.getSelect(0).within(() => cy.get('input').should('be.checked'));
         });
@@ -40,7 +40,7 @@ describe('delete', () => {
         it('select, sort, new row is not selected', () => {
             DashTable.getSelect(0).within(() => cy.get('input').click());
             DashTable.getSelect(0).within(() => cy.get('input').should('be.checked'));
-            cy.get('tr th.column-2 .sort').click({ force: true });
+            cy.get('tr th.column-0 .sort').click({ force: true });
             DashTable.getSelect(0).within(() => cy.get('input').should('not.be.checked'));
         });
     });